Move ModelPatcher to model_patcher.py
